more fixes!
